requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.3
distlib==0.3.1
Django==3.1.1
djangorestframework==3.12.1
filelock==3.0.12
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient @ file:///E:/Brayan/ITP/Noveno%20Semestre/Programacion%20Avanzada/Django/gestiondetareas/mysqlclient-1.4.6-cp38-cp38-win32.whl
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
virtualenv==20.0.32
wrapt==1.12.1
